+++ olive/model/handler/mixin/__init__.py
@@ -1,0 +1,23 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+# --------------------------------------------------------------------------
+from olive.model.handler.mixin.composite import CompositeMixin
+from olive.model.handler.mixin.dummy_inputs import DummyInputsMixin
+from olive.model.handler.mixin.hf_config import HfConfigMixin
+from olive.model.handler.mixin.io_config import IoConfigMixin
+from olive.model.handler.mixin.json import JsonMixin
+from olive.model.handler.mixin.onnx_ep import OnnxEpValidateMixin
+from olive.model.handler.mixin.onnx_graph import OnnxGraphMixin
+from olive.model.handler.mixin.resource import ResourceMixin
+
+__all__ = [
+    "CompositeMixin",
+    "DummyInputsMixin",
+    "HfConfigMixin",
+    "IoConfigMixin",
+    "JsonMixin",
+    "OnnxEpValidateMixin",
+    "OnnxGraphMixin",
+    "ResourceMixin",
+]
