--- olive/model/mixin/__init__.py
@@ -1,23 +1,0 @@
-# -------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-# --------------------------------------------------------------------------
-from olive.model.mixin.composite import CompositeMixin
-from olive.model.mixin.dummy_inputs import DummyInputsMixin
-from olive.model.mixin.hf_config import HfConfigMixin
-from olive.model.mixin.io_config import IoConfigMixin
-from olive.model.mixin.json import JsonMixin
-from olive.model.mixin.onnx_ep import OnnxEpValidateMixin
-from olive.model.mixin.onnx_graph import OnnxGraphMixin
-from olive.model.mixin.resource import ResourceMixin
-
-__all__ = [
-    "CompositeMixin",
-    "DummyInputsMixin",
-    "HfConfigMixin",
-    "IoConfigMixin",
-    "JsonMixin",
-    "OnnxEpValidateMixin",
-    "OnnxGraphMixin",
-    "ResourceMixin",
-]
